Added top-level response and request key maps
Necessary to access these constants via future singleton methods that will be able to mutate the request and response headers. This _inverts the key map_ so symbols are the keys and strings are the values in order to optimize by symbol lookup.

Each of the data objects were refactored accordingly.

Milestone: minor